refactor:  enhance implementation with ApiManager integration
