+++ lib/core/base_states/base_states.dart
@@ -1,6 +1,10 @@
-sealed class BaseStates {}
 
-class InitialState extends BaseStates {}
+
+sealed class BaseStates  {}
+
+class InitialState extends BaseStates {
+
+}
 
 class LoadingState extends BaseStates {}
