+++ lib/core/services/network_srevices.dart
@@ -2,8 +2,8 @@ import 'package:dio/dio.dart';
 import 'package:injectable/injectable.dart';
 import 'package:online_exam_app/core/Constant/app_constant.dart';
 
-@injectable
-class NetworkServices {
+@singleton
+class  NetworkServices {
   late Dio dio;
   NetworkServices() {
     init();
