+++ lib/main.dart
@@ -2,14 +2,16 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:online_exam_app/core/routes/app_routes.dart';
 import 'package:online_exam_app/core/routes/app_routes_generator.dart';
+import 'package:online_exam_app/core/services/api_manager.dart';
 import 'package:online_exam_app/core/theme/app_theme.dart';
 
 import 'core/di/di.dart';
 import 'core/services/shared_prefs.dart';
 
-String? token;
 
-void main() async {
+String? token ;
+
+void main() async{
   WidgetsFlutterBinding.ensureInitialized();
   configureDependencies();
   token = await SharedPrefs().getString("token");
@@ -22,7 +24,6 @@ class MyApp extends StatelessWidget {
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
-    print(token);
     return ScreenUtilInit(
       designSize: const Size(375, 812),
       minTextAdapt: true,
@@ -31,8 +32,7 @@ class MyApp extends StatelessWidget {
         return MaterialApp(
           theme: AppTheme.appTheme,
           debugShowCheckedModeBanner: false,
-          initialRoute:
-              token != null ? AppRoutes.layoutScreen : AppRoutes.loginPage,
+          initialRoute:  token != null ? AppRoutes.homeScreen : AppRoutes.loginPage,
           onGenerateRoute: AppRoutesGenerator.generateRoute,
         );
       },
